moved method to interactive
