+++ interactive.py
@@ -8,7 +8,6 @@ from scipy.spatial.distance import cdist
 from sklearn.metrics import silhouette_score, silhouette_samples
 
 import cmeans as Fuzzy
-from app import scale_score
 
 # confusion matrix
 def computeX2(attrVals, clusters, data, N):
@@ -175,3 +174,15 @@ def icluster(data, terms, userFeedbackTerm, k, userU=-1):
     
     # clusterDocs = [ast.literal_eval(x) for x in clusterDocs]
     return clusterDocs, clusterKeyterms, keyterms, silhouette_avg, scores
+
+def scale_score(avg):
+    negative = False
+    if avg < 0:
+        negative = True
+        avg *= -1
+    score = (50 / (pow(10, (1.0 / 3)))) * pow(10.0 * avg, (1.0 / 3))
+    if negative:
+        score = 50 - score
+    else:
+        score = 50 + score
+    return score
